[FEAT] textFieldUsecase 생성 (#592)
